vercel-requirements.txt
# Flask and extensions
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2

# Database
SQLAlchemy==2.0.40
alembic==1.15.2
greenlet==3.2.1

# Supabase integration
supabase==2.3.0
psycopg2-binary==2.9.9

# Utils
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Mako==1.3.10
typing_extensions==4.13.2
python-dotenv==1.0.1
email-validator==2.1.0.post1
WTForms==3.1.2
Pillow==10.3.0
python-dateutil==2.8.2
requests==2.31.0
pytz==2023.3
cryptography==42.0.2
python-jose==3.3.0
passlib==1.7.4

# APScheduler for challenge scheduling
apscheduler==3.10.4

# Deployment
gunicorn==21.2.0
